refactor(frontend): AuthPopup components
