+++ apps/competition-manager-web/src/Components/AuthPopup/SignIn.tsx
@@ -7,7 +7,7 @@ import { Alert, Box, Button, Divider, Link, Typography } from "@mui/material";
 import { useUserToken } from "../../GlobalsStates/userToken";
 import { login } from "../../utils/auth";
 
-import { Field } from "./Field";
+import { TextFieldWithSchema as Field } from "../../Components/TextFieldWithSchema";
 
 type SignInProps = {
     onToggle: () => void;
@@ -66,17 +66,19 @@ export const SignIn: React.FC<SignInProps> = ({ onToggle }) => {
             >
                 <Field 
                     id="email" 
-                    label="Email" 
+                    label={{ value: 'Email', hasExtrenLabel: true }}
                     value={{ value: email, onChange: setEmail }} 
                     validator={{ Schema$: UserEmail$, isValid: isEmailValid, setIsValid: setIsEmailValid }} 
+                    required
                 />
 
                 <Field 
                     id="password" 
-                    label="Password" 
-                    isPassword 
+                    label={{ value: 'Password', hasExtrenLabel: true }}
                     value={{ value: password, onChange: setPassword }} 
                     validator={{ Schema$: UserPassword$, isValid: isPasswordValid, setIsValid: setIsPasswordValid }}
+                    isPassword
+                    required
                 />
 
                 <Button
